updating scope service and graphql defs for next steps

+++ lib/service/scope-service.ts
@@ -1,4 +1,4 @@
-import { AccessRule, Client, ClientScopeRel, ObjectSearchResultItem, Scope, SearchResultType, Tenant, TenantAvailableScope } from "@/graphql/generated/graphql-types";
+import { AccessRule, Client, ClientScopeRel, ObjectSearchResultItem, Scope, ScopeFilterCriteria, SearchResultType, Tenant, TenantAvailableScope } from "@/graphql/generated/graphql-types";
 import { OIDCContext } from "@/graphql/graphql-context";
 import { GraphQLError } from "graphql/error/GraphQLError";
 import ScopeDao from "../dao/scope-dao";
@@ -7,7 +7,7 @@ import TenantDao from "../dao/tenant-dao";
 import ClientDao from "../dao/client-dao";
 import AccessRuleDao from "../dao/access-rule-dao";
 import { DaoFactory } from "../data-sources/dao-factory";
-import { SCOPE_USE_APPLICATION_MANAGEMENT, SCOPE_USE_DISPLAY, SCOPE_USE_IAM_MANAGEMENT, SEARCH_INDEX_OBJECT_SEARCH } from "@/utils/consts";
+import { ROOT_TENANT_EXCLUSIVE_INTERNAL_SCOPE_NAMES, SCOPE_USE_APPLICATION_MANAGEMENT, SCOPE_USE_DISPLAY, SCOPE_USE_IAM_MANAGEMENT, SEARCH_INDEX_OBJECT_SEARCH, TENANT_TYPE_ROOT_TENANT } from "@/utils/consts";
 import { getOpenSearchClient } from "../data-sources/search";
 
 const scopeDao: ScopeDao = DaoFactory.getInstance().getScopeDao();
@@ -25,8 +25,71 @@ class ScopeService {
     }
 
     
-    public async getScope(tenantId?: string): Promise<Array<Scope>> {
-        return scopeDao.getScope(tenantId);        
+    public async getScope(tenantId?: string, filterBy?: ScopeFilterCriteria): Promise<Array<Scope>> {
+
+        // TODO
+        // Set a tenantId based on the user's oidc context value and use it to query the 
+        // DB. 
+        if(!tenantId){
+            throw new GraphQLError("ERROR_EXPECTING_TENANT_ID_ARGUMENT");
+        }
+
+        let isRootTenant: boolean = false;
+        if(tenantId){
+            const tenant: Tenant | null = await tenantDao.getTenantById(tenantId);
+            if(tenant && tenant.tenantType === TENANT_TYPE_ROOT_TENANT){
+                isRootTenant = true;
+            }
+        }       
+
+        // root tenant + undefined filterBy = Get ALL scope values
+        if(isRootTenant && !filterBy){
+            const arr: Array<Scope> = await scopeDao.getScope();
+            return arr;
+        }
+        if(isRootTenant && filterBy === ScopeFilterCriteria.Existing){
+            const arr: Array<Scope> = await scopeDao.getScope(tenantId);
+            return arr;
+        }
+        if(isRootTenant && filterBy === ScopeFilterCriteria.Available){
+            const arr: Array<Scope> = await scopeDao.getScope();
+            return arr;
+            // const used: Array<Scope> = await scopeDao.getScope(tenantId);
+            // Now get the difference between the two and return it
+            // return arr.filter(
+            //     (s: Scope) => {
+            //         const e: Scope | undefined = used.find(
+            //             (s1: Scope) => s1.scopeId === s.scopeId
+            //         );                    
+            //         return e === undefined;
+            //     }
+            // )
+        }
+        if(!isRootTenant && !filterBy){
+            const arr: Array<Scope> = await scopeDao.getScope(tenantId);
+            return arr;
+        }
+        if(!isRootTenant && filterBy === ScopeFilterCriteria.Existing){
+            const arr: Array<Scope> = await scopeDao.getScope(tenantId);
+            return arr;
+        }
+        if(!isRootTenant && filterBy === ScopeFilterCriteria.Available){
+            const arr: Array<Scope> = await scopeDao.getScope();
+            // Now remove all of the scope values that non-root tenants
+            // do not have access to.
+            return arr.filter(
+                (scope: Scope) => {
+                    const s: string | undefined = ROOT_TENANT_EXCLUSIVE_INTERNAL_SCOPE_NAMES.find(
+                        (s1: string) => {
+                            return s1 === scope.scopeName;
+                        }
+                    )
+                    return s === undefined;
+                }
+            )
+        }
+        return [];
+
     }
 
     public async getScopeById(scopeId: string): Promise<Scope | null> {
@@ -130,6 +193,19 @@ class ScopeService {
 
 
     public async removeScopeFromTenant(tenantId: string, scopeId: string): Promise<void> {
+        const tenant: Tenant | null = await tenantDao.getTenantById(tenantId);
+        if(!tenant){
+            throw new GraphQLError("ERROR_TENANT_DOES_NOT_EXIST");
+        }
+        const scope: Scope | null = await scopeDao.getScopeById(scopeId);
+        if(!scope){
+            throw new GraphQLError("ERROR_SCOPE_DOES_NOT_EXIST");
+        }
+
+        if(tenant.tenantType === TENANT_TYPE_ROOT_TENANT && scope.scopeUse === SCOPE_USE_IAM_MANAGEMENT){
+            throw new GraphQLError("ERROR_CANNOT_REMOVE_IAM_MANAGEMENT_SCOPE_FROM_ROOT_TENANT")
+        }
+        console.log("checkpoint 4");
         return scopeDao.removeScopeFromTenant(tenantId, scopeId);
     }
